+++ Configuration/TCA/Overrides/fe_users.php
@@ -37,67 +37,11 @@ defined('TYPO3') or die('Access denied.');
                 'maxitems' => 1,
             ],
         ],
-        'tx_seminars_auxiliary_records_pid' => [
-            'exclude' => 1,
-            'label' => 'LLL:EXT:seminars/Resources/Private/Language/locallang_db.xlf:fe_groups.tx_seminars_auxiliary_records_pid',
-            'config' => [
-                'type' => 'group',
-                'internal_type' => 'db',
-                'allowed' => 'pages',
-                'default' => 0,
-                'size' => 1,
-                'minitems' => 0,
-                'maxitems' => 1,
-            ],
-        ],
-        'tx_seminars_default_categories' => [
-            'exclude' => 1,
-            'label' => 'LLL:EXT:seminars/Resources/Private/Language/locallang_db.xlf:fe_groups.tx_seminars_default_categories',
-            'config' => [
-                'type' => 'select',
-                'renderType' => 'selectMultipleSideBySide',
-                'foreign_table' => 'tx_seminars_categories',
-                'size' => 10,
-                'minitems' => 0,
-                'maxitems' => 999,
-                'MM' => 'tx_seminars_usergroups_categories_mm',
-                'wizards' => [
-                    'list' => [
-                        'type' => 'popup',
-                        'title' => 'List entries',
-                        'icon' => 'actions-system-list-open',
-                        'params' => [
-                            'table' => 'tx_seminars_categories',
-                            'pid' => '###CURRENT_PID###',
-                        ],
-                        'module' => [
-                            'name' => 'wizard_list',
-                        ],
-                        'JSopenParams' => 'height=480,width=640,status=0,menubar=0,scrollbars=1',
-                    ],
-                ],
-            ],
-        ],
-        'tx_seminars_default_organizer' => [
-            'exclude' => 1,
-            'label' => 'LLL:EXT:seminars/Resources/Private/Language/locallang_db.xlf:fe_groups.tx_seminars_default_organizer',
-            'config' => [
-                'type' => 'select',
-                'renderType' => 'selectSingle',
-                'foreign_table' => 'tx_seminars_organizers',
-                'default' => 0,
-                'size' => 1,
-                'minitems' => 0,
-                'maxitems' => 1,
-            ],
-        ],
     ]
 );
 
 \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addToAllTCAtypes(
     'fe_groups',
     '--div--;LLL:EXT:seminars/Resources/Private/Language/locallang_db.xlf:fe_groups.tab_event_management,' .
-    'tx_seminars_events_pid,' .
-    'tx_seminars_auxiliary_records_pid,' .
-    'tx_seminars_default_categories, tx_seminars_default_organizer'
+    'tx_seminars_events_pid'
 );