+++ Tests/LegacyUnit/Model/FrontEndUserGroupTest.php
@@ -4,10 +4,6 @@ declare(strict_types=1);
 
 namespace OliverKlee\Seminars\Tests\LegacyUnit\Model;
 
-use OliverKlee\Oelib\DataStructures\Collection;
-use OliverKlee\Oelib\Mapper\MapperRegistry;
-use OliverKlee\Seminars\Mapper\CategoryMapper;
-use OliverKlee\Seminars\Mapper\OrganizerMapper;
 use OliverKlee\Seminars\Model\FrontEndUserGroup;
 use PHPUnit\Framework\TestCase;
 
@@ -24,64 +20,6 @@ final class FrontEndUserGroupTest extends TestCase
     protected function setUp(): void
     {
         $this->subject = new FrontEndUserGroup();
-    }
-
-    ///////////////////////////////////////////////
-    // Tests concerning getAuxiliaryRecordsPid().
-    ///////////////////////////////////////////////
-
-    /**
-     * @test
-     */
-    public function getAuxiliaryRecordsPidWithoutPidReturnsZero(): void
-    {
-        $this->subject->setData([]);
-
-        self::assertEquals(
-            0,
-            $this->subject->getAuxiliaryRecordsPid()
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function getAuxiliaryRecordsPidWithPidReturnsPid(): void
-    {
-        $this->subject->setData(['tx_seminars_auxiliary_records_pid' => 42]);
-
-        self::assertEquals(
-            42,
-            $this->subject->getAuxiliaryRecordsPid()
-        );
-    }
-
-    ///////////////////////////////////////////////
-    // Tests concerning hasAuxiliaryRecordsPid().
-    ///////////////////////////////////////////////
-
-    /**
-     * @test
-     */
-    public function hasAuxiliaryRecordsPidWithoutPidReturnsFalse(): void
-    {
-        $this->subject->setData([]);
-
-        self::assertFalse(
-            $this->subject->hasAuxiliaryRecordsPid()
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function hasAuxiliaryRecordsPidWithPidReturnsTrue(): void
-    {
-        $this->subject->setData(['tx_seminars_auxiliary_records_pid' => 42]);
-
-        self::assertTrue(
-            $this->subject->hasAuxiliaryRecordsPid()
-        );
     }
 
     //////////////////////////////////////////////////
@@ -135,111 +73,6 @@ final class FrontEndUserGroupTest extends TestCase
         self::assertEquals(
             42,
             $this->subject->getEventRecordPid()
-        );
-    }
-
-    //////////////////////////////////////////
-    // Tests concerning getDefaultCategories
-    //////////////////////////////////////////
-
-    /**
-     * @test
-     */
-    public function getDefaultCategoriesForNoCategoriesReturnsAList(): void
-    {
-        $this->subject->setData(['tx_seminars_default_categories' => new Collection()]);
-
-        self::assertInstanceOf(Collection::class, $this->subject->getDefaultCategories());
-    }
-
-    /**
-     * @test
-     */
-    public function getDefaultCategoriesForOneAssignedCategoryReturnsThisCategoryInList(): void
-    {
-        $list = new Collection();
-        $category = MapperRegistry::get(CategoryMapper::class)->getNewGhost();
-
-        $list->add($category);
-        $this->subject->setData(['tx_seminars_default_categories' => $list]);
-
-        self::assertSame(
-            $category,
-            $this->subject->getDefaultCategories()->first()
-        );
-    }
-
-    //////////////////////////////////////////
-    // Tests concerning hasDefaultCategories
-    //////////////////////////////////////////
-
-    /**
-     * @test
-     */
-    public function hasDefaultCategoriesForNoAssignedCategoriesReturnsFalse(): void
-    {
-        $this->subject->setData(['tx_seminars_default_categories' => new Collection()]);
-
-        self::assertFalse(
-            $this->subject->hasDefaultCategories()
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function hasDefaultCategoriesForOneAssignedCategoryReturnsTrue(): void
-    {
-        $list = new Collection();
-        $list->add(MapperRegistry::get(CategoryMapper::class)->getNewGhost());
-
-        $this->subject->setData(['tx_seminars_default_categories' => $list]);
-
-        self::assertTrue(
-            $this->subject->hasDefaultCategories()
-        );
-    }
-
-    /////////////////////////////////////////////////////////////////
-    // Tests concerning getDefaultOrganizer and hasDefaultOrganizer
-    /////////////////////////////////////////////////////////////////
-
-    /**
-     * @test
-     */
-    public function getDefaultOrganizerForSetOrganizerReturnsIt(): void
-    {
-        $organizer = MapperRegistry::get(OrganizerMapper::class)->getNewGhost();
-        $this->subject->setData(['tx_seminars_default_organizer' => $organizer]);
-
-        self::assertSame(
-            $organizer,
-            $this->subject->getDefaultOrganizer()
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function hasDefaultOrganizerForSetOrganizerReturnsTrue(): void
-    {
-        $organizer = MapperRegistry::get(OrganizerMapper::class)->getNewGhost();
-        $this->subject->setData(['tx_seminars_default_organizer' => $organizer]);
-
-        self::assertTrue(
-            $this->subject->hasDefaultOrganizer()
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function hasDefaultOrganizerForNotSetOrganizerReturnsFalse(): void
-    {
-        $this->subject->setData(['tx_seminars_default_organizer' => null]);
-
-        self::assertFalse(
-            $this->subject->hasDefaultOrganizer()
         );
     }
 }