+++ ext_tables.sql
@@ -10,10 +10,7 @@ CREATE TABLE fe_users (
 # Table structure for table 'fe_groups'
 #
 CREATE TABLE fe_groups (
-    tx_seminars_events_pid int(11) unsigned DEFAULT '0' NOT NULL,
-    tx_seminars_auxiliary_records_pid int(11) unsigned DEFAULT '0' NOT NULL,
-    tx_seminars_default_categories int(11) unsigned DEFAULT '0' NOT NULL,
-    tx_seminars_default_organizer int(11) unsigned DEFAULT '0' NOT NULL
+    tx_seminars_events_pid int(11) unsigned DEFAULT '0' NOT NULL
 );
 
 #